refactor: 클라이언트가 구입 금액을 입력한다
