+++ README.md
@@ -22,7 +22,7 @@
 ## 🚀 2단계 - 로또(자동)
 
 ### TODO
-- [ ] 클라이언트가 구입 금액을 입력한다. - InputReader
+- [x] 클라이언트가 구입 금액을 입력한다. - InputReader
 - [ ] 클라이언트가 입력한 값은 숫자이여야 한다. - InputReader
 - [ ] 구입 금액에 로또 가격을 나눠서 몇장을 구매했는지 알려줘야한다.
 - [ ] 구매한 로또의 숫자를 출력한다.
