+++ src/main/kotlin/LottoRunner.kt
@@ -1,0 +1,8 @@
+import view.InputView
+
+class LottoRunner(private val inputView: InputView) {
+
+    fun startLotto() {
+        inputView.enterTheMoney()
+    }
+}
