+++ src/main/kotlin/view/InputReader.kt
@@ -1,4 +1,4 @@
-package util
+package view
 
 fun interface InputReader {
     fun raedLine(): String
